requirements.txt
qtpy
numpy
scipy
lmfit
ponyorm
pyqtgraph
